Move file related function to utils.py.
